Fix max allowed code words calculation
The code was checking that the total number of code words in the data
region, including error correction codes was not larger than 928.
Instead, the length descriptor, which does not include the error
correction code words, should be under 928.

Therefore it's possible to encode more data than was allowed.

issue #9