Fix byte compaction for chunks with low values
When encoding a 6-byte chunk, the byte compaction must yield exactly
5 codewords, in order to properly decode to the same 6 bytes.

The `to_base()` util function had no notion of the expected number of
result digits, so it couldn't add the necessary leading 0s when the
base-900 encoded value was small enough to fit in 4 "digits" (codewords)
This can happen when the leftmost input byte is 0, and the next one is low.

This would lead to off-by-one situations and the codewords would be
unexpectedly shifted by one, ending with an incorrect PDF417 rendering.

Extended `test_compaction` to cover this case.